Split dependee and dependent lookup into separate classes
